Rework platformer into Castlevania-style multiplayer race
Complete overhaul of the platformer game mode:

Game logic (breakpoint-platformer):
- Enemy system: Skeleton (patrol), Bat (sine-wave), Knight (2HP patrol),
  Medusa (float + projectiles) with per-type AI tick functions
- Melee combat: whip attack with hitbox, active frames, cooldown, damage,
  invincibility frames, death/respawn with time penalty
- Course generation: 300x30 tile grid, 15 rooms with branching paths
  (upper=safer/longer, lower=more enemies/shorter), room templates
  (Corridor, Staircase, TowerClimb, CathedralHall, CryptDepths, etc.)
- Rubber banding: leader gets +50% enemies, last place gets better items
- 7 power-ups: HolyWater, Crucifix, SpeedBoots, DoubleJump, ArmorUp,
  Invincibility, WhipExtend with Mario Kart-style tiered selection
- Race-only mode at 20Hz tick rate, "Castlevania Rush" name
- Compact u8 Tile serialization to stay under 64KB protocol limit

Client rendering (breakpoint-client):
- Sprite/texture pipeline: new sprite.vert/frag shaders, texture atlas
  support with NEAREST filtering, Quad mesh, Sprite material type
- Sprite atlas module with named regions, animation support, 41 sprites
- Camera: tighter Z=-15 framing, 3-unit lead in race direction
- Gothic theme: dark stone, blood-red spikes, gold finish
- 5 new audio events: Attack, Hit, Death, EnemyKill, Checkpoint
- HUD: HP hearts, death counter, active power-up, race position

674 tests passing, 0 clippy warnings.

+++ crates/breakpoint-client/src/bridge.rs
@@ -217,7 +217,7 @@ fn build_platformer_hud(app: &App) -> serde_json::Value {
         .unwrap_or_default();
 
     // Race position: rank by furthest X progress among non-eliminated players
-    let mut positions: Vec<(u32, f32)> = state
+    let mut positions: Vec<(u64, f32)> = state
         .players
         .iter()
         .filter(|(_, p)| !p.eliminated)
